Moved Effects to own files, cached constants
